+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Crevice Changelog
 
 ## Unreleased Changes
+* Added `AsStd140::std140_size` for easily pre-sizing buffers.
 
 ## 0.3.0 (2020-09-22)
 * Added `Std140::as_bytes`, reducing the need to work with bytemuck directly.